+++ mandelbulber2/src/opencl_engine_render_dof_phase2.cpp
@@ -344,10 +344,10 @@ bool cOpenClEngineRenderDOFPhase2::WriteBuffersToQueue()
 	return true;
 }
 
-bool cOpenClEngineRenderDOFPhase2::ProcessQueue(size_t pixelsLeft, size_t pixelIndex)
+bool cOpenClEngineRenderDOFPhase2::ProcessQueue(qint64 pixelsLeft, qint64 pixelIndex)
 {
-	size_t limitedWorkgroupSize = optimalJob.workGroupSize;
-	size_t stepSize = optimalJob.stepSize;
+	qint64 limitedWorkgroupSize = optimalJob.workGroupSize;
+	qint64 stepSize = optimalJob.stepSize;
 
 	if (optimalJob.stepSize > pixelsLeft)
 	{
@@ -448,7 +448,7 @@ bool cOpenClEngineRenderDOFPhase2::Render(
 		// requires initialization for all opencl devices
 		// requires optimalJob for all opencl devices
 
-		for (size_t pixelIndex = 0; pixelIndex < width * height; pixelIndex += optimalJob.stepSize)
+		for (qint64 pixelIndex = 0; pixelIndex < width * height; pixelIndex += optimalJob.stepSize)
 		{
 			size_t pixelsLeft = width * height - pixelIndex;
