+++ mandelbulber2/src/opencl_engine_render_ssao.cpp
@@ -259,10 +259,10 @@ bool cOpenClEngineRenderSSAO::WriteBuffersToQueue()
 	return true;
 }
 
-bool cOpenClEngineRenderSSAO::ProcessQueue(size_t pixelsLeft, size_t pixelIndex)
+bool cOpenClEngineRenderSSAO::ProcessQueue(qint64 pixelsLeft, qint64 pixelIndex)
 {
-	size_t limitedWorkgroupSize = optimalJob.workGroupSize;
-	size_t stepSize = optimalJob.stepSize;
+	qint64 limitedWorkgroupSize = optimalJob.workGroupSize;
+	qint64 stepSize = optimalJob.stepSize;
 
 	if (optimalJob.stepSize > pixelsLeft)
 	{
@@ -354,7 +354,7 @@ bool cOpenClEngineRenderSSAO::Render(cImage *image, bool *stopRequest)
 		// insert device for loop here
 		// requires initialization for all opencl devices
 		// requires optimalJob for all opencl devices
-		for (size_t pixelIndex = 0; pixelIndex < width * height; pixelIndex += optimalJob.stepSize)
+		for (qint64 pixelIndex = 0; pixelIndex < width * height; pixelIndex += optimalJob.stepSize)
 		{
 			size_t pixelsLeft = width * height - pixelIndex;
 			UpdateOptimalJobStart(pixelsLeft);
